HAI-2018 Reorganise usercontact and email extraction funtionality to more suitable packages and files.
